Remove unecessary indirection added by SmithyBaseTask class
