+++ smithy-base/src/main/java/software/amazon/smithy/gradle/tasks/SmithyFormatTask.java
@@ -7,6 +7,7 @@ package software.amazon.smithy.gradle.tasks;
 
 import java.io.File;
 import javax.inject.Inject;
+import org.gradle.StartParameter;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.tasks.TaskAction;
 import software.amazon.smithy.utils.ListUtils;
@@ -28,8 +29,8 @@ public abstract class SmithyFormatTask extends AbstractSmithyCliTask {
     private static final String DESCRIPTION = "Formats smithy models.";
 
     @Inject
-    public SmithyFormatTask(ObjectFactory objectFactory) {
-        super(objectFactory);
+    public SmithyFormatTask(ObjectFactory objectFactory, StartParameter startParameter) {
+        super(objectFactory, startParameter);
         setDescription(DESCRIPTION);
     }
